requirements.txt
zeep>=4.1.0
xmltodict>=0.12.0
cryptography>=40.0.0
logtail>=1.0.0
python-dotenv>=0.21.0
